pool: Fix some issues in `PUT /farmer` (#272)
* pool: Missing failure response + respond with bool, not with puzzle hash

Responding with the puzzle hash is not correct according to the spec.

* pool: Fix type of `authentication_public_key` in the "update dict"

We can't put a `G1Element` here because the `from_json_dict` expects a 
hex string and raises.

* pool: Fix difficulty updates and update the others values immediately

`add_farmer_record` doesn't even adjust the difficulty in its SQL query 
so we use `update_difficulty` now.

This also leads to the other auth key and payout instructions to become 
updated instant and only the difficulty update being delayed. But still, 
the request is locked until the difficulty update went through.